+++ rust/src/detect/transforms/domain.rs
@@ -15,14 +15,11 @@
  * 02110-1301, USA.
  */
 
-use super::{
-    InspectionBufferCheckAndExpand, InspectionBufferLength, InspectionBufferPtr,
-    InspectionBufferTruncate,
-};
 use crate::detect::SIGMATCH_NOOPT;
 use suricata_sys::sys::{
     DetectEngineCtx, DetectEngineThreadCtx, InspectionBuffer, SCDetectHelperTransformRegister,
-    SCDetectSignatureAddTransform, SCTransformTableElmt, Signature,
+    SCDetectSignatureAddTransform, SCTransformTableElmt, Signature, SCInspectionBufferCheckAndExpand,
+    SCInspectionBufferTruncate,
 };
 
 use std::os::raw::{c_int, c_void};
@@ -50,14 +47,14 @@ fn get_domain(input: &[u8], output: &mut [u8]) -> u32 {
 unsafe extern "C" fn domain_transform(
     _det: *mut DetectEngineThreadCtx, buffer: *mut InspectionBuffer, _ctx: *mut c_void,
 ) {
-    let input = InspectionBufferPtr(buffer);
-    let input_len = InspectionBufferLength(buffer);
+    let input = (*buffer).inspect;
+    let input_len = (*buffer).inspect_len;
     if input.is_null() || input_len == 0 {
         return;
     }
     let input = build_slice!(input, input_len as usize);
 
-    let output = InspectionBufferCheckAndExpand(buffer, input_len);
+    let output = SCInspectionBufferCheckAndExpand(buffer, input_len);
     if output.is_null() {
         // allocation failure
         return;
@@ -66,7 +63,7 @@ unsafe extern "C" fn domain_transform(
 
     let output_len = get_domain(input, output);
 
-    InspectionBufferTruncate(buffer, output_len);
+    SCInspectionBufferTruncate(buffer, output_len);
 }
 
 unsafe extern "C" fn tld_setup(
@@ -89,14 +86,14 @@ fn get_tld(input: &[u8], output: &mut [u8]) -> u32 {
 unsafe extern "C" fn tld_transform(
     _det: *mut DetectEngineThreadCtx, buffer: *mut InspectionBuffer, _ctx: *mut c_void,
 ) {
-    let input = InspectionBufferPtr(buffer);
-    let input_len = InspectionBufferLength(buffer);
+    let input = (*buffer).inspect;
+    let input_len = (*buffer).inspect_len;
     if input.is_null() || input_len == 0 {
         return;
     }
     let input = build_slice!(input, input_len as usize);
 
-    let output = InspectionBufferCheckAndExpand(buffer, input_len);
+    let output = SCInspectionBufferCheckAndExpand(buffer, input_len);
     if output.is_null() {
         // allocation failure
         return;
@@ -105,7 +102,7 @@ unsafe extern "C" fn tld_transform(
 
     let output_len = get_tld(input, output);
 
-    InspectionBufferTruncate(buffer, output_len);
+    SCInspectionBufferTruncate(buffer, output_len);
 }
 
 #[no_mangle]
