+++ rust/src/detect/transforms/dotprefix.rs
@@ -15,14 +15,11 @@
  * 02110-1301, USA.
  */
 
-use super::{
-    InspectionBufferCheckAndExpand, InspectionBufferLength, InspectionBufferPtr,
-    InspectionBufferTruncate,
-};
 use crate::detect::SIGMATCH_NOOPT;
 use suricata_sys::sys::{
     DetectEngineCtx, DetectEngineThreadCtx, InspectionBuffer, SCDetectHelperTransformRegister,
-    SCDetectSignatureAddTransform, SCTransformTableElmt, Signature,
+    SCDetectSignatureAddTransform, SCTransformTableElmt, Signature, SCInspectionBufferCheckAndExpand,
+    SCInspectionBufferTruncate,
 };
 
 use std::os::raw::{c_int, c_void};
@@ -48,17 +45,17 @@ fn dot_prefix_transform_do(input: &[u8], output: &mut [u8]) {
 unsafe extern "C" fn dot_prefix_transform(
     _det: *mut DetectEngineThreadCtx, buffer: *mut InspectionBuffer, _ctx: *mut c_void,
 ) {
-    let input_len = InspectionBufferLength(buffer);
+    let input_len = (*buffer).inspect_len;
     if input_len == 0 {
         return;
     }
-    let output = InspectionBufferCheckAndExpand(buffer, input_len + 1);
+    let output = SCInspectionBufferCheckAndExpand(buffer, input_len + 1);
     if output.is_null() {
         // allocation failure
         return;
     }
     // get input after possible realloc
-    let input = InspectionBufferPtr(buffer);
+    let input = (*buffer).inspect;
     if input.is_null() {
         // allocation failure
         return;
@@ -68,7 +65,7 @@ unsafe extern "C" fn dot_prefix_transform(
 
     dot_prefix_transform_do(input, output);
 
-    InspectionBufferTruncate(buffer, input_len + 1);
+    SCInspectionBufferTruncate(buffer, input_len + 1);
 }
 
 #[no_mangle]
