+++ src/Makefile.am
@@ -136,6 +136,7 @@ noinst_HEADERS = \
 	detect-engine-file.h \
 	detect-engine-frame.h \
 	detect-engine-helper.h \
+	detect-engine-inspect-buffer.h \
 	detect-engine-iponly.h \
 	detect-engine-loader.h \
 	detect-engine-mpm.h \
@@ -724,6 +725,7 @@ libsuricata_c_a_SOURCES = \
 	detect-engine-file.c \
 	detect-engine-frame.c \
 	detect-engine-helper.c \
+	detect-engine-inspect-buffer.c \
 	detect-engine-iponly.c \
 	detect-engine-loader.c \
 	detect-engine-mpm.c \
