+++ src/bindgen.h
@@ -33,6 +33,9 @@
 
 #include "app-layer-protos.h"
 #include "suricata-plugin.h"
+// do not export struct fields only used for debug validation
+// do this after suricata-plugin.h which needs autoconf.h to define SC_PACKAGE_VERSION
+#undef DEBUG_VALIDATION
 #include "output-eve-bindgen.h"
 #include "detect-engine-register.h"
 #include "detect-engine-buffer.h"
